requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.3.6
redis==5.0.1
structlog==25.4.0
pytest==8.2.0
httpx==0.27.0
python-multipart==0.0.9
alembic==1.13.1